make the sentinel conversions explicit (C4245, C4018)
Every signed-to-unsigned conversion left in modorganizer turned out to be
a deliberate sentinel or an already-guarded comparison, not a bug -- but
each one relied on implicit wrapping to say so:

  NoSelection = -1                     (std::size_t)
  firstRowIndex = -1                   (unsigned int)
  TreeItem::childIndex returns -1      (std::size_t, "not a child")
  TreeItem() : TreeItem(nullptr, -1)   (unsigned int index)
  m_exitCode(-1)                       (DWORD)

They now cast explicitly, so the intent is in the code rather than in the
conversion rules. instancemanagerdialog also passes the existing
NoSelection constant instead of a bare -1, which is what it meant.

modlistview.cpp compared a signed modIndex against the unsigned
getNumMods(). The preceding `modIndex < 0` check already made this safe,
so the cast is explicit rather than corrective.

C4245, C4018 and C4189 are now at zero. 0 errors.

+++ src/instancemanagerdialog.h
@@ -87,7 +87,7 @@ protected:
   void showEvent(QShowEvent* e) override;
 
 private:
-  static const std::size_t NoSelection = -1;
+  static const std::size_t NoSelection = static_cast<std::size_t>(-1);
 
   std::unique_ptr<Ui::InstanceManagerDialog> ui;
   PluginContainer& m_pc;
